+++ app/components/keyset/Listing.vue
@@ -28,12 +28,6 @@
     </template>
 
     <template #body>
-      <!-- <UPageHeader :title="title">
-        <template #description>
-          <SharedPageHeaderDescription :description="description" />
-        </template>
-      </UPageHeader> -->
-
       <UPageGrid
         v-if="keysets.length"
         class="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-3 2xl:grid-cols-4 3xl:grid-cols-6 4xl:grid-cols-6 gap-4"